Move C++ compilation smoke tests to module
