+++ tests/cpp_compile/mod.rs
@@ -8,6 +8,8 @@ use proc_macro2::TokenStream;
 use std::borrow::Cow;
 use std::path::{Path, PathBuf};
 
+mod smoke_test;
+
 /// Helper for setting up a test that:
 ///
 /// 1. Takes a `#[cxx::bridge]` and generates `.cc` and `.h` files,
@@ -186,68 +188,5 @@ impl CompilationResult {
         let single_error_line = error_lines.into_iter().next().unwrap();
         eprintln!("Got single error as expected: {single_error_line}");
         single_error_line
-    }
-}
-
-mod test {
-    use super::Test;
-    use quote::quote;
-
-    #[test]
-    fn test_success_smoke_test() {
-        let test = Test::new(quote! {
-            #[cxx::bridge]
-            mod ffi {
-                unsafe extern "C++" {
-                    include!("include.h");
-                    pub fn do_cpp_thing();
-                }
-            }
-        });
-        test.write_file("include.h", "void do_cpp_thing();");
-        test.compile().assert_success();
-    }
-
-    #[test]
-    fn test_failure_smoke_test() {
-        let test = Test::new(quote! {
-            #[cxx::bridge]
-            mod ffi {
-                unsafe extern "C++" {
-                    include!("include.h");
-                }
-            }
-        });
-        test.write_file(
-            "include.h",
-            r#"
-                static_assert(false, "This is a failure smoke test");
-            "#,
-        );
-        let err_msg = test.compile().expect_single_error();
-        assert!(err_msg.contains("This is a failure smoke test"));
-    }
-
-    #[test]
-    #[should_panic = "Unexpectedly more than 1 error line was present"]
-    fn test_failure_with_unexpected_extra_error_line() {
-        let test = Test::new(quote! {
-            #[cxx::bridge]
-            mod ffi {
-                unsafe extern "C++" {
-                    include!("include.h");
-                }
-            }
-        });
-        test.write_file(
-            "include.h",
-            r#"
-                static_assert(false, "First error line");
-                static_assert(false, "Second error line");
-            "#,
-        );
-
-        // We `should_panic` inside `expect_single_error` below:
-        let _ = test.compile().expect_single_error();
     }
 }
